event_queue: split kafka facade into facade package modules
