+++ components/bot_detector/event_queue/facade/__init__.py
@@ -1,31 +1,40 @@
-from .factory import QueueFactory
-from .facade import (
-    DataToPredictConsumer,
-    DataToPredictProducer,
-    DataToPredictQueue,
+from bot_detector.event_queue.structs import (
     DataToPredictStruct,
     HighScoreStruct,
     NotFoundStruct,
-    PlayersNotFoundConsumer,
-    PlayersNotFoundProducer,
-    PlayersNotFoundQueue,
-    PlayersScrapedConsumer,
-    PlayersScrapedProducer,
-    PlayersScrapedQueue,
-    PlayersToScrapeConsumer,
-    PlayersToScrapeProducer,
-    PlayersToScrapeQueue,
-    ReportsToInsertConsumer,
-    ReportsToInsertProducer,
-    ReportsToInsertQueue,
     ReportsToInsertStruct,
     ScrapedStruct,
-    Settings,
     ToScrapeStruct,
 )
 
+from .data_to_predict import (
+    DataToPredictConsumer,
+    DataToPredictProducer,
+    DataToPredictQueue,
+)
+from .players_not_found import (
+    PlayersNotFoundConsumer,
+    PlayersNotFoundProducer,
+    PlayersNotFoundQueue,
+)
+from .players_scraped import (
+    PlayersScrapedConsumer,
+    PlayersScrapedProducer,
+    PlayersScrapedQueue,
+)
+from .players_to_scrape import (
+    PlayersToScrapeConsumer,
+    PlayersToScrapeProducer,
+    PlayersToScrapeQueue,
+)
+from .reports_to_insert import (
+    ReportsToInsertConsumer,
+    ReportsToInsertProducer,
+    ReportsToInsertQueue,
+)
+from .settings import Settings
+
 __all__ = [
-    "QueueFactory",
     "Settings",
     "ToScrapeStruct",
     "ScrapedStruct",
