+++ components/bot_detector/event_queue/facade/settings.py
@@ -1,0 +1,6 @@
+from pydantic import Field
+from pydantic_settings import BaseSettings
+
+
+class Settings(BaseSettings):
+    KAFKA_BOOTSTRAP_SERVERS: str = Field(default=...)
